requirements.txt
bzip2=1.0.6=vc14_3
cycler=0.10.0=py35_0
freetype=2.5.5=vc14_2
icu=57.1=vc14_0
jpeg=9b=vc14_0
libpng=1.6.27=vc14_0
libtiff=4.0.6=vc14_3
lxml=3.8.0=py35_0
matplotlib=2.0.2=np113py35_0
mkl=2017.0.1=0
numpy=1.13.1=py35_0
olefile=0.44=py35_0
openssl=1.0.2l=vc14_0
pillow=4.1.1=py35_0
pip=9.0.1=py35_1
pyparsing=2.1.4=py35_0
pyqt=5.6.0=py35_2
python=3.5.3=3
python-dateutil=2.6.0=py35_0
pytz=2017.2=py35_0
qt=5.6.2=vc14_4
scikit-learn=0.18.2=np113py35_0
scipy=0.19.1=np113py35_0
setuptools=27.2.0=py35_1
sip=4.18=py35_0
six=1.10.0=py35_0
tk=8.5.18=vc14_0
vs2015_runtime=14.0.25420=0
wheel=0.29.0=py35_0
zlib=1.2.8=vc14_3

backports.weakref==1.0rc1
bleach==1.5.0
click==6.7
easydict==1.7
flask==0.12.2
html5lib==0.9999999
itsdangerous==0.24
jinja2==2.9.6
markdown==2.2.0
markupsafe==1.0
opencv-python==3.2.0.7
protobuf==3.3.0
sqlalchemy==1.1.12
tensorflow-gpu==1.2.0
werkzeug==0.12.2